requirements.txt
PyQt5
pyopengl
matplotlib
reportlab
pyyaml
customtkinter
numpy
